chore: Add comments

+++ src/main/java/org/javaalgorithminterview/ch18_binary_search/P75_2_Immigration.java
@@ -11,10 +11,11 @@ public class P75_2_Immigration {
         long answer = 0;
 
         long left = 1;
-        long right = Arrays.stream(times).max().getAsInt() * n;
+        long right = (long) Arrays.stream(times).max().getAsInt() * n;
 
         long mid = right;
 
+        // Why should this comparison operation contain equal(=)?
         while (left <= right) {
             long calcN = 0;
 
@@ -22,13 +23,14 @@ public class P75_2_Immigration {
                 calcN += (mid / time);
             }
 
+            // The answer continue to converge until the remainder is zero when dividing the total time by a unit time
             if (calcN >= n) {
                 answer = mid;
                 right = mid - 1;
             } else {
                 left = mid + 1;
             }
-
+            // Why isn't the logic to calculate mid at the beginning?
             mid = left + (right - left) / 2;
         }
         return answer;
